requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==13.9.4
six==1.17.0
typing_extensions==4.12.2
yt-dlp==2025.1.15
